print an error message for empty runnable binary wrapper
As a follow up to https://github.com/matt-sm/rules_foreign_cc/commit/1d8e362a22be9ead657205185df2d1544283b81d, we are still experiencing intermittent occurances of this issue whereby the wrapper script is empty on windows, and have been unable to reliably repro. This error message will at least give the user a clue of what the actual problem is.

+++ foreign_cc/utils.bzl
@@ -50,6 +50,10 @@ def runnable_binary(name, binary, foreign_cc_target, match_binary_name = False, 
 
     wrapper_cmd = """
     sed s@EXECUTABLE@$(rlocationpath {name})@g $(location @rules_foreign_cc//foreign_cc/private:runnable_binary_wrapper.sh) > tmp
+    if [[ ! -s tmp ]]; then
+        echo "ERROR: runnable binary wrapper script is empty" >&2
+        exit 1
+    fi
     cp tmp $@
     """
